requirements2.txt
alembic==0.9.3
appdirs==1.4.4
bidict==0.21.2
bleach==2.0.0
blinker==1.4
cachelib==0.1.1
click==8.0.0
distlib==0.3.2
dominate==2.3.1
Faker==10.0.0
filelock==3.0.12
Flask==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-Reuploaded==1.2.0
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
Flask-WTF==1.0.0
greenlet==1.1.0
gunicorn==19.7.1
html5lib==0.999999999
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==2.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.5
python-dateutil==2.6.1
python-dotenv==0.18.0
python-editor==1.0.3
python-engineio==4.2.0
python-socketio==5.3.0
six==1.10.0
SQLAlchemy==1.3.24
text-unidecode==1.3
virtualenv==20.0.31
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.0.0
WTForms==2.1
